Allow update profile without class info

+++ backend/src/controllers/UpdateController.ts
@@ -26,10 +26,17 @@ export default class UpdateController {
         } = req.body;
 
         let { schedule } = req.body;
-        schedule = JSON.parse(schedule);
+
+        if (schedule)
+            schedule = JSON.parse(schedule);
 
         const user_id = req.userId;
-        const avatar = req.file.filename;
+        
+        let avatar = null;
+
+        if (req.file)
+            avatar = req.file.filename;
+
         let oldAvatar = undefined;
 
         //console.log({ name, surname, email, whatsapp, bio, subject, cost, schedule, user_id, avatar})
@@ -51,47 +58,60 @@ export default class UpdateController {
         const trx = await db.transaction();
 
         try {
-            await trx('users').where('id', user_id).update({
-                name,
-                surname,
-                email,
-                whatsapp,
-                bio,
-                avatar,
-            });
-
-            const classExists = await trx('classes').where('subject', subject).first();
-            let classesIds;
-
-            if (classExists) {
-                classesIds = await trx('classes').where('subject', subject).update({
-                    subject,
-                    cost,
-                    user_id,
-               });  
+            if (avatar) {
+                await trx('users').where('id', user_id).update({
+                    name,
+                    surname,
+                    email,
+                    whatsapp,
+                    bio,
+                    avatar,
+                });
             }else {
-                classesIds = await trx('classes').insert({
-                    subject,
-                    cost,
-                    user_id,
-               });
-               classesIds = classesIds[0];
+                await trx('users').where('id', user_id).update({
+                    name,
+                    surname,
+                    email,
+                    whatsapp,
+                    bio,
+                });
             }
 
-            const class_id = classesIds;
+            if (subject) {
+                const classExists = await trx('classes').where('subject', subject).first();
+            
+                let classesIds;
+    
+                if (classExists) {
+                    classesIds = await trx('classes').where('subject', subject).update({
+                        subject,
+                        cost,
+                        user_id,
+                   });  
+                }else {
+                    classesIds = await trx('classes').insert({
+                        subject,
+                        cost,
+                        user_id,
+                   });
+                   classesIds = classesIds[0];
+                }
+    
+                const class_id = classesIds;
+    
+                const classSchedule = schedule.map((scheduleItem: ScheduleItem) => {
+                    return {
+                        class_id,
+                        week_day: scheduleItem.week_day,
+                        from: convertHourToMinutes(scheduleItem.from),
+                        to: convertHourToMinutes(scheduleItem.to),
+                    };
+                });
+    
+                await trx('class_schedule').where('class_id', '=', class_id).delete();
+                await trx('class_schedule').insert(classSchedule);
+            }
 
-            const classSchedule = schedule.map((scheduleItem: ScheduleItem) => {
-                return {
-                    class_id,
-                    week_day: scheduleItem.week_day,
-                    from: convertHourToMinutes(scheduleItem.from),
-                    to: convertHourToMinutes(scheduleItem.to),
-                };
-            });
-
-            await trx('class_schedule').where('class_id', '=', class_id).delete();
-            await trx('class_schedule').insert(classSchedule);
-         
             await trx.commit();
 
             res.send();
